feat(company-url): refactored classes to their own files
